refactor(app-paths): extract uc-app-paths as the single directory-layout authority
ADR-008 P5-0c. P5-0 left app-data-root resolution duplicated between uc-platform
and uc-daemon-process (the thin crate cannot depend on heavy uc-platform). Extract
the path-resolution policy into a new thin crate uc-app-paths (deps: dirs + std
only): APP_DIR_NAME, profile suffix, portable redirect, data/cache base resolution.

uc-platform now delegates to it (keeps the AppDirsPort impl, AppDirsError mapping,
test override, default_profile/dev-profile feature, and a thin resolve_profile
wrapper shared with secure_storage). portable.rs becomes a re-export shim.
uc-daemon-process depends on uc-app-paths and deletes its duplicated app_data_root.rs.

Single source of truth -> drift between the daemon's and client's path resolution
is now impossible at compile time. Future cache/log/user-data dir splits change one
place (uc-app-paths). Zero behavior change: resolved data_root/cache_root are
byte-identical for both consumers; uc-platform still requires data+cache, daemon-
process still needs only data (the benign P5-0 divergence preserved).

cargo check --workspace clean (incl. --features dev-profile); uc-app-paths tree =
dirs only; daemon-client still free of iroh/diesel; lib tests pass (uc-app-paths 5,
uc-platform 67, uc-daemon-process 11, uc-daemon-local 25). Also removes uc-platform's
now-unused direct dirs dep.
